+++ src/libs/ajax/ajax-common.ts
@@ -13,30 +13,6 @@ export const withAppIdentifier =
     return wrappedFetch(url, _.merge(options, appIdentifier));
   };
 
-export type RequesterPaysErrorInfo = {
-  requesterPaysError: boolean;
-};
-
-export const isRequesterPaysErrorInfo = (error: any): error is RequesterPaysErrorInfo => {
-  return error != null && typeof error === 'object' && 'requesterPaysError' in error;
-};
-
-export const checkRequesterPaysError = async (response): Promise<RequesterPaysErrorInfo> => {
-  if (response.status === 400) {
-    const data = await response.text();
-    const requesterPaysErrorInfo: RequesterPaysErrorInfo = {
-      requesterPaysError: responseContainsRequesterPaysError(data),
-    };
-    return Object.assign(new Response(new Blob([data]), response), requesterPaysErrorInfo);
-  }
-  const unrecognizedErrorInfo: RequesterPaysErrorInfo = { requesterPaysError: false };
-  return Object.assign(response, unrecognizedErrorInfo);
-};
-
-export const responseContainsRequesterPaysError = (responseText) => {
-  return _.includes('requester pays', responseText);
-};
-
 export const fetchLeo = _.flow(
   withUrlPrefix(`${getConfig().leoUrlRoot}/`),
   withRetryAfterReloadingExpiredAuthToken
