+++ src/libs/ajax/goggle-storage-models.ts
@@ -1,0 +1,7 @@
+export type RequesterPaysErrorInfo = {
+  requesterPaysError: boolean;
+};
+
+export const isRequesterPaysErrorInfo = (error: any): error is RequesterPaysErrorInfo => {
+  return error != null && typeof error === 'object' && 'requesterPaysError' in error;
+};
