+++ src/workspaces/common/requester-pays/bucket-utils.js
@@ -2,11 +2,11 @@ import { abandonedPromise } from '@terra-ui-packages/core-utils';
 import _ from 'lodash/fp';
 import { useState } from 'react';
 import { h } from 'react-hyperscript-helpers';
-import RequesterPaysModal from 'src/components/RequesterPaysModal';
-import { isRequesterPaysErrorInfo } from 'src/libs/ajax/ajax-common';
+import { isRequesterPaysErrorInfo } from 'src/libs/ajax/goggle-storage-models';
 import { forwardRefWithName } from 'src/libs/react-utils';
 import { requesterPaysProjectStore } from 'src/libs/state';
 import * as Utils from 'src/libs/utils';
+import { RequesterPaysModal } from 'src/workspaces/common/requester-pays/RequesterPaysModal';
 
 // Returns true if the error object passed in was marked as a requester pays error. Note that this parsing of the
 // error response is done in checkRequesterPaysError, which is used by the GoogleStorage ajax library.
