+++ src/workspaces/common/state/useWorkspace.ts
@@ -5,9 +5,8 @@ import { defaultLocation } from 'src/analysis/utils/runtime-utils';
 import { Link } from 'src/components/common';
 import { locationTypes } from 'src/components/region-common';
 import { Ajax } from 'src/libs/ajax';
-import { responseContainsRequesterPaysError } from 'src/libs/ajax/ajax-common';
 import { AzureStorage } from 'src/libs/ajax/AzureStorage';
-import { saToken } from 'src/libs/ajax/GoogleStorage';
+import { responseContainsRequesterPaysError, saToken } from 'src/libs/ajax/GoogleStorage';
 import { ErrorCallback, withErrorHandling, withErrorIgnoring, withErrorReporting } from 'src/libs/error';
 import Events, { extractWorkspaceDetails } from 'src/libs/events';
 import { clearNotification, notify } from 'src/libs/notifications';
